CR: moved exceptions mappers from package `mappers` to `exceptions`
