+++ services-api/src/main/java/io/scalecube/services/exceptions/ServiceClientErrorMapper.java
@@ -1,4 +1,4 @@
-package io.scalecube.services.exceptions.mappers;
+package io.scalecube.services.exceptions;
 
 import io.scalecube.services.api.ServiceMessage;
