+++ services-api/src/main/java/io/scalecube/services/methods/ServiceMethodRegistry.java
@@ -1,6 +1,6 @@
 package io.scalecube.services.methods;
 
-import io.scalecube.services.exceptions.mappers.ServiceProviderErrorMapper;
+import io.scalecube.services.exceptions.ServiceProviderErrorMapper;
 
 public interface ServiceMethodRegistry {
