+++ services-gateway-http/src/main/java/io/scalecube/services/gateway/http/HttpGatewayAcceptor.java
@@ -17,7 +17,7 @@ import io.scalecube.services.api.ErrorData;
 import io.scalecube.services.api.Qualifier;
 import io.scalecube.services.api.ServiceMessage;
 import io.scalecube.services.api.ServiceMessage.Builder;
-import io.scalecube.services.exceptions.mappers.DefaultErrorMapper;
+import io.scalecube.services.exceptions.DefaultErrorMapper;
 import io.scalecube.services.gateway.GatewayMetrics;
 import io.scalecube.services.gateway.ReferenceCountUtil;
 import io.scalecube.services.transport.api.DataCodec;
