+++ services-gateway-websocket/src/main/java/io/scalecube/services/gateway/ws/WebsocketGatewayAcceptor.java
@@ -2,7 +2,7 @@ package io.scalecube.services.gateway.ws;
 
 import io.scalecube.services.ServiceCall;
 import io.scalecube.services.api.ServiceMessage;
-import io.scalecube.services.exceptions.mappers.DefaultErrorMapper;
+import io.scalecube.services.exceptions.DefaultErrorMapper;
 import io.scalecube.services.gateway.GatewayMetrics;
 import io.scalecube.services.gateway.ReferenceCountUtil;
 import io.scalecube.services.gateway.ws.GatewayMessage.Builder;
