+++ services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/ExtendedNioEventLoopGroup.java
@@ -65,9 +65,7 @@ public class ExtendedNioEventLoopGroup extends NioEventLoopGroup {
   @Override
   public ChannelFuture register(Channel channel) {
     EventLoop eventExecutor = eventLoopChooser.apply(channel, iterator());
-    return eventExecutor != null
-        ? eventExecutor.register(channel)
-        : super.register(channel);
+    return eventExecutor != null ? eventExecutor.register(channel) : super.register(channel);
   }
 
   @Override
