+++ services/src/main/java/io/scalecube/services/methods/ServiceMethodRegistryImpl.java
@@ -1,7 +1,7 @@
 package io.scalecube.services.methods;
 
 import io.scalecube.services.Reflect;
-import io.scalecube.services.exceptions.mappers.ServiceProviderErrorMapper;
+import io.scalecube.services.exceptions.ServiceProviderErrorMapper;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
